+++ src/objects/trefoil.js
@@ -1,5 +1,6 @@
 import * as THREE from 'three';
 
+// Trefoil curve: parametric family with p=2, q=3 by default ama orjinali istediğim gibi olmadı aslında
 export class TrefoilCurve extends THREE.Curve {
   constructor(a = 2.0, b = 1.0, p = 2, q = 3) {
     super();
@@ -15,7 +16,7 @@ export class TrefoilCurve extends THREE.Curve {
     const b = this.b;
     const p = this.p;
     const q = this.q;
-
+//denklemimiz bu
     const x = (a + b * Math.cos(q * theta)) * Math.cos(p * theta);
     const y = (a + b * Math.cos(q * theta)) * Math.sin(p * theta);
     const z = b * Math.sin(q * theta);